refactor!: remove deprecated GeneratedVaadinCheckboxGroup
