--- vaadin-checkbox-flow-parent/vaadin-checkbox-flow/src/main/java/com/vaadin/flow/component/checkbox/GeneratedVaadinCheckboxGroup.java
@@ -1,486 +1,0 @@
-/*
- * Copyright 2000-2022 Vaadin Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.vaadin.flow.component.checkbox;
-
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import com.vaadin.flow.component.AbstractSinglePropertyField;
-import com.vaadin.flow.component.ComponentEvent;
-import com.vaadin.flow.component.ComponentEventListener;
-import com.vaadin.flow.component.HasStyle;
-import com.vaadin.flow.component.HasTheme;
-import com.vaadin.flow.component.NotSupported;
-import com.vaadin.flow.component.Tag;
-import com.vaadin.flow.component.dependency.JsModule;
-import com.vaadin.flow.component.dependency.NpmPackage;
-import com.vaadin.flow.function.SerializableBiFunction;
-import com.vaadin.flow.function.SerializableFunction;
-import com.vaadin.flow.shared.Registration;
-
-/**
- * <p>
- * Description copied from corresponding location in WebComponent:
- * </p>
- * <p>
- * {@code <vaadin-checkbox-group>} is a Polymer element for grouping
- * vaadin-checkboxes.
- * </p>
- * <p>
- * &lt;vaadin-checkbox-group label=&quot;Preferred language of
- * contact:&quot;&gt; &lt;vaadin-checkbox
- * value=&quot;en&quot;&gt;English&lt;/vaadin-checkbox&gt; &lt;vaadin-checkbox
- * value=&quot;fr&quot;&gt;Français&lt;/vaadin-checkbox&gt; &lt;vaadin-checkbox
- * value=&quot;de&quot;&gt;Deutsch&lt;/vaadin-checkbox&gt;
- * &lt;/vaadin-checkbox-group&gt;
- * </p>
- * <h3>Styling</h3>
- * <p>
- * The following shadow DOM parts are available for styling:
- * </p>
- * <table>
- * <thead>
- * <tr>
- * <th>Part name</th>
- * <th>Description</th>
- * </tr>
- * </thead> <tbody>
- * <tr>
- * <td>{@code label}</td>
- * <td>The label element</td>
- * </tr>
- * <tr>
- * <td>{@code group-field}</td>
- * <td>The element that wraps checkboxes</td>
- * </tr>
- * <tr>
- * <td>{@code error-message}</td>
- * <td>The error message element</td>
- * </tr>
- * </tbody>
- * </table>
- * <p>
- * The following state attributes are available for styling:
- * </p>
- * <table>
- * <thead>
- * <tr>
- * <th>Attribute</th>
- * <th>Description</th>
- * <th>Part name</th>
- * </tr>
- * </thead> <tbody>
- * <tr>
- * <td>{@code disabled}</td>
- * <td>Set when the checkbox group and its children are disabled.</td>
- * <td>:host</td>
- * </tr>
- * <tr>
- * <td>{@code has-label}</td>
- * <td>Set when the element has a label</td>
- * <td>:host</td>
- * </tr>
- * <tr>
- * <td>{@code has-value}</td>
- * <td>Set when the element has a value</td>
- * <td>:host</td>
- * </tr>
- * <tr>
- * <td>{@code required}</td>
- * <td>Set when the element is required</td>
- * <td>:host</td>
- * </tr>
- * <tr>
- * <td>{@code invalid}</td>
- * <td>Set when the element is invalid</td>
- * <td>:host</td>
- * </tr>
- * </tbody>
- * </table>
- * <p>
- * See
- * <a href="https://github.com/vaadin/vaadin-themable-mixin/wiki">ThemableMixin
- * – how to apply styles for shadow parts</a>
- * </p>
- */
-@Tag("vaadin-checkbox-group")
-@NpmPackage(value = "@vaadin/polymer-legacy-adapter", version = "24.0.0-alpha6")
-@JsModule("@vaadin/polymer-legacy-adapter/style-modules.js")
-@NpmPackage(value = "@vaadin/checkbox-group", version = "24.0.0-alpha6")
-@JsModule("@vaadin/checkbox-group/src/vaadin-checkbox-group.js")
-public abstract class GeneratedVaadinCheckboxGroup<R extends GeneratedVaadinCheckboxGroup<R, T>, T>
-        extends AbstractSinglePropertyField<R, T>
-        implements HasStyle, HasTheme {
-
-    /**
-     * Adds theme variants to the component.
-     *
-     * @param variants
-     *            theme variants to add
-     */
-    public void addThemeVariants(CheckboxGroupVariant... variants) {
-        getThemeNames().addAll(
-                Stream.of(variants).map(CheckboxGroupVariant::getVariantName)
-                        .collect(Collectors.toList()));
-    }
-
-    /**
-     * Removes theme variants from the component.
-     *
-     * @param variants
-     *            theme variants to remove
-     */
-    public void removeThemeVariants(CheckboxGroupVariant... variants) {
-        getThemeNames().removeAll(
-                Stream.of(variants).map(CheckboxGroupVariant::getVariantName)
-                        .collect(Collectors.toList()));
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * The current disabled state of the checkbox group. True if group and all
-     * internal checkboxes are disabled.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code disabled} property from the webcomponent
-     */
-    protected boolean isDisabledBoolean() {
-        return getElement().getProperty("disabled", false);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * The current disabled state of the checkbox group. True if group and all
-     * internal checkboxes are disabled.
-     * </p>
-     *
-     * @param disabled
-     *            the boolean value to set
-     */
-    protected void setDisabled(boolean disabled) {
-        getElement().setProperty("disabled", disabled);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * String used for the label element.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code label} property from the webcomponent
-     */
-    protected String getLabelString() {
-        return getElement().getProperty("label");
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * String used for the label element.
-     * </p>
-     *
-     * @param label
-     *            the String value to set
-     */
-    protected void setLabel(String label) {
-        getElement().setProperty("label", label == null ? "" : label);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Error to show when the input value is invalid.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code errorMessage} property from the webcomponent
-     */
-    protected String getErrorMessageString() {
-        return getElement().getProperty("errorMessage");
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Error to show when the input value is invalid.
-     * </p>
-     *
-     * @param errorMessage
-     *            the String value to set
-     */
-    protected void setErrorMessage(String errorMessage) {
-        getElement().setProperty("errorMessage",
-                errorMessage == null ? "" : errorMessage);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Specifies that the user must fill in a value.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code required} property from the webcomponent
-     */
-    protected boolean isRequiredBoolean() {
-        return getElement().getProperty("required", false);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Specifies that the user must fill in a value.
-     * </p>
-     *
-     * @param required
-     *            the boolean value to set
-     */
-    protected void setRequired(boolean required) {
-        getElement().setProperty("required", required);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * This property is set to true when the control value is invalid.
-     *
-     * @return the {@code invalid} property from the webcomponent
-     */
-    protected boolean isInvalidBoolean() {
-        return getElement().getProperty("invalid", false);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * This property is set to true when the control value is invalid.
-     * </p>
-     *
-     * @param invalid
-     *            the boolean value to set
-     */
-    protected void setInvalid(boolean invalid) {
-        getElement().setProperty("invalid", invalid);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Returns true if {@code value} is valid. {@code <iron-form>} uses this to
-     * check the validity or all its elements.
-     * </p>
-     * <p>
-     * This function is not supported by Flow because it returns a
-     * <code>boolean</code>. Functions with return types different than void are
-     * not supported at this moment.
-     */
-    @NotSupported
-    protected void validate() {
-    }
-
-    public static class InvalidChangeEvent<R extends GeneratedVaadinCheckboxGroup<R, ?>>
-            extends ComponentEvent<R> {
-        private final boolean invalid;
-
-        public InvalidChangeEvent(R source, boolean fromClient) {
-            super(source, fromClient);
-            this.invalid = source.isInvalidBoolean();
-        }
-
-        public boolean isInvalid() {
-            return invalid;
-        }
-    }
-
-    /**
-     * Adds a listener for {@code invalid-changed} events fired by the
-     * webcomponent.
-     *
-     * @param listener
-     *            the listener
-     * @return a {@link Registration} for removing the event listener
-     */
-    protected Registration addInvalidChangeListener(
-            ComponentEventListener<InvalidChangeEvent<R>> listener) {
-        return getElement()
-                .addPropertyChangeListener("invalid",
-                        event -> listener.onComponentEvent(
-                                new InvalidChangeEvent<R>((R) this,
-                                        event.isUserOriginated())));
-    }
-
-    /**
-     * Constructs a new GeneratedVaadinCheckboxGroup component with the given
-     * arguments.
-     *
-     * @param initialValue
-     *            the initial value to set to the value
-     * @param defaultValue
-     *            the default value to use if the value isn't defined
-     * @param elementPropertyType
-     *            the type of the element property
-     * @param presentationToModel
-     *            a function that converts a string value to a model value
-     * @param modelToPresentation
-     *            a function that converts a model value to a string value
-     * @param <P>
-     *            the property type
-     */
-    public <P> GeneratedVaadinCheckboxGroup(T initialValue, T defaultValue,
-            Class<P> elementPropertyType,
-            SerializableFunction<P, T> presentationToModel,
-            SerializableFunction<T, P> modelToPresentation) {
-        super("value", defaultValue, elementPropertyType, presentationToModel,
-                modelToPresentation);
-        if (initialValue != null) {
-            setPresentationValue(initialValue);
-        }
-    }
-
-    /**
-     * Constructs a new GeneratedVaadinCheckboxGroup component with the given
-     * arguments.
-     *
-     * @param initialValue
-     *            the initial value to set to the value
-     * @param defaultValue
-     *            the default value to use if the value isn't defined
-     * @param acceptNullValues
-     *            whether <code>null</code> is accepted as a model value
-     */
-    public GeneratedVaadinCheckboxGroup(T initialValue, T defaultValue,
-            boolean acceptNullValues) {
-        super("value", defaultValue, acceptNullValues);
-        if (initialValue != null) {
-            setPresentationValue(initialValue);
-        }
-    }
-
-    /**
-     * Constructs a new GeneratedVaadinCheckboxGroup component with the given
-     * arguments.
-     * <p>
-     * If {@code isInitialValueOptional} is {@code true} then the initial value
-     * is used only if element has no {@code "value"} property value, otherwise
-     * element {@code "value"} property is ignored and the initial value is set.
-     *
-     * @param initialValue
-     *            the initial value to set to the value
-     * @param defaultValue
-     *            the default value to use if the value isn't defined
-     * @param elementPropertyType
-     *            the type of the element property
-     * @param presentationToModel
-     *            a function that accepts this component and a property value
-     *            and returns a model value
-     * @param modelToPresentation
-     *            a function that accepts this component and a model value and
-     *            returns a property value
-     * @param isInitialValueOptional
-     *            if {@code isInitialValueOptional} is {@code true} then the
-     *            initial value is used only if element has no {@code "value"}
-     *            property value, otherwise element {@code "value"} property is
-     *            ignored and the initial value is set
-     * @param <P>
-     *            the property type
-     */
-    public <P> GeneratedVaadinCheckboxGroup(T initialValue, T defaultValue,
-            Class<P> elementPropertyType,
-            SerializableBiFunction<R, P, T> presentationToModel,
-            SerializableBiFunction<R, T, P> modelToPresentation,
-            boolean isInitialValueOptional) {
-        super("value", defaultValue, elementPropertyType, presentationToModel,
-                modelToPresentation);
-        if ((getElement().getProperty("value") == null
-                || !isInitialValueOptional) && initialValue != null) {
-            setPresentationValue(initialValue);
-        }
-    }
-
-    /**
-     * Constructs a new GeneratedVaadinCheckboxGroup component with the given
-     * arguments.
-     * <p>
-     * If {@code isInitialValueOptional} is {@code true} then the initial value
-     * is used only if element has no {@code "value"} property value, otherwise
-     * element {@code "value"} property is ignored and the initial value is set.
-     *
-     * @param initialValue
-     *            the initial value to set to the value
-     * @param defaultValue
-     *            the default value to use if the value isn't defined
-     * @param elementPropertyType
-     *            the type of the element property
-     * @param presentationToModel
-     *            a function that accepts this component and a property value
-     *            and returns a model value
-     * @param modelToPresentation
-     *            a function that accepts this component and a model value and
-     *            returns a property value
-     * @param <P>
-     *            the property type
-     */
-    public <P> GeneratedVaadinCheckboxGroup(T initialValue, T defaultValue,
-            Class<P> elementPropertyType,
-            SerializableBiFunction<R, P, T> presentationToModel,
-            SerializableBiFunction<R, T, P> modelToPresentation) {
-        this(initialValue, defaultValue, elementPropertyType,
-                presentationToModel, modelToPresentation, false);
-    }
-
-    /**
-     * Default constructor.
-     */
-    public GeneratedVaadinCheckboxGroup() {
-        this(null, null, null, (SerializableFunction) null,
-                (SerializableFunction) null);
-    }
-
-}
